2.3.0

+++ src/main/java/me/starlitembers/emberslib/Util.java
@@ -1,7 +1,7 @@
 package me.starlitembers.emberslib;
 
 public class Util {
-    public static int timeToTicks(int hours, int minutes, int seconds, int ticks){
-        return (((hours * 60) * 60) * 20) + ((minutes * 60) * 20) + (seconds * 20) + ticks;
+    public static long timeToTicks(int hours, int minutes, int seconds, int ticks){
+        return (((hours * 60L) * 60) * 20) + ((minutes * 60L) * 20) + (seconds * 20L) + ticks;
     }
 }
